requirements.txt
astroid==1.6.6
autopep8==1.5.4
backports.functools-lru-cache==1.6.1
configparser==4.0.2
enum34==1.1.10
futures==3.3.0
isort==4.3.21
lazy-object-proxy==1.5.1
mccabe==0.6.1
pycodestyle==2.6.0
pylint==1.9.5
python-dotenv==0.14.0
singledispatch==3.4.0.3
six==1.15.0
toml==0.10.1
typing==3.7.4.3
wrapt==1.12.1
